implementations/design_04_hybrid/requirements.txt
chromadb>=0.4.0
pyyaml>=6.0
httpx>=0.24.0
